requirements.txt
openai>=1.0.0
pydantic>=2.0.0 
gspread
google-auth
msgpack-rpc-python 
numpy
matplotlib
lxml
mako
msgpack
glob2
tornado 
backports.zoneinfo; python_version < "3.9" 
backports.ssl_match_hostname 